convert device id to a long

+++ api/src/main/java/xyz/klinker/messenger/api/service/DeviceService.java
@@ -31,7 +31,7 @@ public interface DeviceService {
     AddDeviceResponse add(@Body AddDeviceRequest request);
 
     @POST("devices/update/{id}")
-    Object update(@Path("id") int id, @Query("account_id") String accountId,
+    Object update(@Path("id") long id, @Query("account_id") String accountId,
                   @Query("name") String name, @Query("fcm_token") String fcmToken);
 
     @POST("devices/remove/{id}")
